try to display generated image on visualizeactivity

+++ app/src/main/java/edu/gatech/wavevis/app/Drawable.java
@@ -1,5 +1,5 @@
 package edu.gatech.wavevis.app;
 
 public interface Drawable {
-    byte[] draw(int width, int height);
+    int[] draw(int width, int height);
 }
