+++ app/src/main/java/edu/gatech/wavevis/app/FFTFrame.java
@@ -31,27 +31,27 @@ public class FFTFrame implements Drawable{
     }
 
     @Override
-    public byte[] draw(int width, int height) {
+    public int[] draw(int width, int height) {
         NVector logFrame = frameData.log10();
         float minLog = logFrame.min();
         float maxLog = logFrame.max();
         float logRange = maxLog - minLog;
-        byte[] imageBuffer = new byte[width * height * 4];
+        int[] imageBuffer = new int[width * height * 4];
         for (int i = 0; i < width; i++) {
             int binIndex = (int) ((float) i * (float) getBinCount() / (float) width);
             int normPower = (int) (((logFrame.getValue(binIndex) - minLog) / logRange) * (float) height);
             for (int j = 0; j < height; j++) {
                 int mapIndex = (j * width + i) * 4; // index of imageBuffer[x, y, 0]
                 if (j <= normPower) {
-                    imageBuffer[mapIndex] = (byte) 255;
-                    imageBuffer[mapIndex + 1] = (byte) 255;
-                    imageBuffer[mapIndex + 2] = (byte) 0;
-                    imageBuffer[mapIndex + 3] = (byte) 0;
+                    imageBuffer[mapIndex] = 255;
+                    imageBuffer[mapIndex + 1] = 255;
+                    imageBuffer[mapIndex + 2] = 0;
+                    imageBuffer[mapIndex + 3] = 0;
                 } else {
-                    imageBuffer[mapIndex] = (byte) 255;
-                    imageBuffer[mapIndex + 1] = (byte) 0;
-                    imageBuffer[mapIndex + 2] = (byte) 0;
-                    imageBuffer[mapIndex + 3] = (byte) 0;
+                    imageBuffer[mapIndex] = 255;
+                    imageBuffer[mapIndex + 1] = 0;
+                    imageBuffer[mapIndex + 2] = 0;
+                    imageBuffer[mapIndex + 3] = 0;
                 }
             }
         }
